・HDL用HSファイル追加 ・HspVarInt64_LtI のミスを修正（toropippiさん、ありがとうございます） ・cfunc64v のみ命令でも呼べるように ・hspint64.as 更新

+++ src/hspvar_int64.cpp
@@ -275,7 +275,7 @@ static void HspVarInt64_GtI( PDAT *pval, const void *val )
 // Lt
 static void HspVarInt64_LtI( PDAT *pval, const void *val )
 {
-	*((int *)pval) = ( *GetPtr(pval) < *((float *)(val)) );
+	*((int *)pval) = ( *GetPtr(pval) < *((INT64 *)(val)) );
 	*aftertype = HSPVAR_FLAG_INT;
 }
